+++ src/dirtyfields/dirtyfields.py
@@ -81,9 +81,9 @@ class DirtyFieldsMixin(object):
             field_value = getattr(self, field.attname)
 
             if isinstance(field_value, File):
-                # Uses the path for files due to a perfomance regression caused by Django 3.1.
+                # Uses the name for files due to a perfomance regression caused by Django 3.1.
                 # For more info see: https://github.com/romgar/django-dirtyfields/issues/165
-                field_value = field_value.path if field_value else ''
+                field_value = field_value.name
 
             # If current field value is an expression, we are not evaluating it
             if isinstance(field_value, (BaseExpression, Combinable)):